refactor: Split where_pushdown.rs into focused modules
Split the 1063-line where_pushdown.rs into focused submodules for better
maintainability and navigation:

- mod.rs: Public API and core types (PredicateCategory, PredicateDecomposition)
- classification.rs: Predicate classification logic
- table_refs.rs: Table reference extraction utilities
- or_conditions.rs: OR condition handling and filter extraction
- tests.rs: Test suite (12 tests, all passing)

Benefits:
- Improved maintainability with smaller, focused modules
- Better testability through isolated concerns
- Easier navigation with clear separation of responsibilities
- Reduced cognitive load (each module handles one aspect)

All existing tests pass. No changes to public API or behavior.

Closes #2385
